* updated timer event code to use float times

+++ engine/timer/timers.h
@@ -1,7 +1,7 @@
 #pragma once
 
 namespace timer {
-	bool addTimer(const string& name, const string& action, unsigned int frequency_ms = 0, unsigned int start_ms = 0);
+	bool addTimer(const string& name, const string& action, float frequency_ms = 0.0f, float start_ms = 0.0f);
 	bool removeTimer(const string& name);
 	void fireTimers();
 }
